+++ railties/lib/rails/deprecator.rb
@@ -2,6 +2,6 @@
 
 module Rails
   def self.deprecator # :nodoc:
-    ActiveSupport::Deprecation.instance
+    @deprecator ||= ActiveSupport::Deprecation.new
   end
 end